+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # PayPal Android SDK Release Notes
 
 ## unreleased
+* PayPalNativeCheckout
+  * Bump native-checkout version to release `1.2.0`
+  * Adding in support for the merchant passed email feature (opt in only)
 
 * Breaking Changes
   * CardPayments